+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Localrep computeplan status field, that is synced when recieving computetask event update instead of in the computeplan view  (#762)
 - Compute and store cp dates and duration during sync (#757)
 - Local representation of model assets (#784)
+- Use localrep data in newsfeed view (#795)
 
 ### Changed
 - When resyncing the local representation, only fetch the latest events (#656)